==> tiKPC_IMC/scripts/02_B_immune_cell_clustering.Rmd <==
---
title: "02_B_immune_cell_clustering"
author: "Jacob Mitchell"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(dplyr)
library(Rphenograph)
library(ggplot2)
library(ComplexHeatmap)
library(dittoSeq)
library(pals)

set.seed(123)
sessionInfo()
```

```{r functions}
plot_imc_violin <- function(df, group.by, feature) {
  pl <- ggplot(df, aes(x = .data[[group.by]], y = .data[[feature]])) +
    geom_violin(aes(fill = .data[[group.by]]), scale = "width") +
    stat_summary(fun = "median", shape = 1, size = 0.3, color = "#000000") +
    scale_fill_manual(values = colorassigned) +
    theme_bw() +
    theme(legend.position = "none")
  
  return(pl)
}
```

```{r}
result_dir <- "processed_data/02_B_immune_cell_clustering"
if(!dir.exists(result_dir)) dir.create(result_dir)
figure_dir <- "figures/02_B_immune_cell_clustering"
if(!dir.exists(figure_dir)) dir.create(figure_dir)
```

```{r}
cell_meta <- read.csv("processed_data/02_A_broad_cell_clustering/cell_meta.csv", row.names = 1)

# Expression values scaled from 0 to 1 across all cells
imc_norm <- readRDS("processed_data/02_A_broad_cell_clustering/IMC_normalized.rds")

# arcsinh-transformed raw values
imc_arcsin <- readRDS("processed_data/02_A_broad_cell_clustering/IMC_arcsinh.rds")

rownames(cell_meta) <- cell_meta$cell
imc <- cbind(cell_meta, imc_norm)
imc$cluster <- factor(imc$cluster, levels = seq(max(cell_meta$cluster)))
```

```{r}
# immune clusters
immune_cluster_ids <- c(
  # "1", "2", "3", "4", "5", "6", "8", "10", "13", "36", "37"
  
  # mean normalized CD45 > 0.4
  # "1", "2", "3", "4", "5", "6", "8", "10", "13", "14", "20", "26", "27", "35", "36", "37", "38"
  
  "3", "4", "5", "6", "7", "9", "12", "13", "16",
  "28", "29", "30", "31", "32"
)
meta_colnames <- c("cell", "cluster", "immune_cluster", "ImageId", "CellId")

immune_imc <- imc[imc$cluster %in% immune_cluster_ids,]
immune_imc_norm <- immune_imc[, !(colnames(immune_imc) %in% meta_colnames)] %>%
  as.matrix()
```

```{r}
set.seed(1234)
phenographout <- Rphenograph(immune_imc_norm, k = 30)
immune_imc$immune_cluster <- factor(membership(phenographout[[2]]))

immune_cell_meta <- immune_imc[, colnames(immune_imc) %in% meta_colnames]
```

```{r}
features <- colnames(immune_imc_norm)

cluster_mean <- immune_imc %>%
  group_by(immune_cluster) %>% 
  dplyr::summarize_all(list(mean))

cluster_mean_mat <- as.matrix(cluster_mean[, features])

rownames(cluster_mean_mat) <- 1:nrow(cluster_mean_mat)

cluster_scaled <- scale(cluster_mean_mat)

rownames(cluster_scaled) <- 1:nrow(cluster_scaled)
```

```{r}
# create cluster summary heatmaps
n_levels <- nlevels(immune_imc$immune_cluster)
ditto_cols <- dittoSeq::dittoColors(1)

# colors for the heatmap
legend_breaks <- seq(from = 0, to = 1, by = 0.2)
colorassigned <- setNames(
  ditto_cols[seq_along(unique(cluster_mean$immune_cluster))],
  sort(unique(cluster_mean$immune_cluster))
)
color_list <- list(clusters=colorassigned)
color_list_byoriginal <- colorassigned[match((cluster_mean$immune_cluster),names(colorassigned))]

rAbar <- rowAnnotation(
  clusters=cluster_mean$immune_cluster,
  col=color_list, gp = gpar(col = "white", lwd = .5),
  counts= anno_barplot(
    as.vector(table(immune_imc$immune_cluster)), gp = gpar(fill=colorassigned),
    border = F, bar_width = 0.75, width = unit(2,"cm")
  )
)

pdf(paste0(figure_dir, "/", "immune_clusterheatmap.pdf"),width=10,height=8)
Heatmap(
  cluster_scaled,
  column_title="Phenograph Clusters", name = "scaled",
  col=rev(brewer.rdbu(100)), cluster_columns = T, cluster_rows = T,
  border = NA, rect_gp = gpar(col = "white", lwd = .5), right_annotation = rAbar,
  show_row_names = T, row_names_gp = gpar(fontsize=7), column_names_gp = gpar(fontsize=10),
  heatmap_legend_param = list(at=seq(from = round(min(cluster_scaled)), to = round(max(cluster_scaled)))),
  width = ncol(cluster_scaled)*unit(4, "mm"), height = nrow(cluster_scaled)*unit(4, "mm")
)
dev.off() 
```

```{r}
# save results
saveRDS(phenographout, file = paste0(result_dir, "/IMC_immune_phenograph.rds"))
saveRDS(immune_imc, file = paste0(result_dir, "/IMC_immune_clustered.rds"))
saveRDS(immune_imc_norm, file = paste0(result_dir, "/IMC_immune_normalized.rds"))

# cell meta data with cluster assignment
write.csv(immune_cell_meta, file = paste0(result_dir, "/immune_cell_meta.csv"))

# matrices used to render heatmaps
write.csv(cluster_mean_mat, file = paste0(result_dir, "/IMCimmunecluster_mean.csv"))
write.csv(cluster_scaled, file = paste0(result_dir, "/IMCimmunecluster_scaledMean.csv"))
```

```{r}
immune_imc_arcsin <- cbind(immune_cell_meta, imc_arcsin[immune_cell_meta$cell,])

cell_features <- colnames(cluster_scaled)
for(i in seq_along(cell_features)){
  feat <- cell_features[i]
  pl <- plot_imc_violin(
    df = immune_imc_arcsin, feature = feat, group.by = "immune_cluster"
  )
  print(pl)
  ggsave(
    plot = pl, filename = paste0(figure_dir, "/", "vln_", "immune_cluster", "_", feat, ".png"),
    dpi = 300, width = unit(8, "in"), height = unit(6, "in")
  )
}
```
