==> tiKPC_IMC/scripts/02_C_cell_type_annotation.Rmd <==
---
title: "02_C_cell_type_annotation"
author: "Jacob Mitchell"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(plyr)
library(dplyr)
library(Rphenograph)
library(ggplot2)
library(ComplexHeatmap)
library(scales)
library(dittoSeq)
library(pals)
library(umap)

set.seed(123)
sessionInfo()
```

```{r}
result_dir <- "processed_data/02_C_cell_type_annotation"
if(!dir.exists(result_dir)) dir.create(result_dir)
figure_dir <- "figures/02_C_cell_type_annotation"
if(!dir.exists(figure_dir)) dir.create(figure_dir)
```

```{r}
# IMC expression values

# Expression values scaled from 0 to 1 across all cells
imc_norm <- readRDS("processed_data/02_A_broad_cell_clustering/IMC_normalized.rds")

# arcsinh-transformed raw values
imc_arcsin <- readRDS("processed_data/02_A_broad_cell_clustering/IMC_arcsinh.rds")
```


```{r}
all_meta <- read.csv(
  "processed_data/02_A_broad_cell_clustering/cell_meta.csv", 
  row.names = 1
)
immune_meta <- read.csv(
  "processed_data/02_B_immune_cell_clustering/immune_cell_meta.csv", 
  row.names = 1
)

cell_meta <- dplyr::left_join(
  x = all_meta, y = immune_meta[, c("cell", "immune_cluster")], by = "cell"
)
```

```{r}
# broad cluster annotation
broad_cluster_ls <- list(
  '1' = "Unspecified",
  '2' = "Epithelial",
  '3' = "Immune",
  '4' = "Immune",
  '5' = "Immune",
  '6' = "Immune",
  '7' = "Immune",
  '8' = "Epithelial",
  '9' = "Immune",
  '10' = "Epithelial",
  '11' = "CAF",
  '12' = "Immune",
  '13' = "Immune",
  '14' = "Epithelial",
  '15' = "Epithelial",
  '16' = "Immune",
  '17' = "Epithelial",
  '18' = "CAF",
  '19' = "Unspecified",
  '20' = "Epithelial",
  '21' = "Unspecified",
  '22' = "CAF",
  '23' = "Epithelial",
  '24' = "Unspecified",
  '25' = "Epithelial",
  '26' = "Unspecified",
  '27' = "Unspecified",
  '28' = "Immune",
  '29' = "Immune",
  '30' = "Immune",
  '31' = "Immune",
  '32' = "Immune"
)

immune_broad_cluster_ls <- list(
  '1' = "Bcell",
  '2' = "Tcell",
  '3' = "Tcell",
  '4' = "Myeloid",
  '5' = "Tcell",
  '6' = "Myeloid",
  '7' = "Myeloid",
  '8' = "Myeloid",
  '9' = "Myeloid",
  '10' = "PDPN+",
  '11' = "Bcell",
  '12' = "Myeloid",
  '13' = "Epithelial",
  '14' = "Bcell",
  '15' = "Myeloid",
  '16' = "Epithelial",
  '17' = "Bcell",
  '18' = "Myeloid",
  '19' = "CD31+",
  '20' = "PDPN+",
  '21' = "Epithelial",
  '22' = "Myeloid",
  '23' = "Myeloid",
  '24' = "Myeloid",
  '25' = "Myeloid",
  '26' = "Epithelial",
  '27' = "Myeloid",
  '28' = "Epithelial",
  '29' = "Bcell"
)

immune_specific_cluster_ls <- list(
  '1' = "Bcell",
  '2' = "CD8_Tcell",
  '3' = "CD4_Th",
  '4' = "Macrophage",
  '5' = "CD4_Treg",
  '6' = "Macrophage",
  '7' = "Macrophage",
  '8' = "Granulocyte",
  '9' = "Granulocyte",
  '10' = "PDPN+",
  '11' = "Bcell",
  '12' = "Macrophage",
  '13' = "Epithelial",
  '14' = "Bcell",
  '15' = "Macrophage",
  '16' = "Epithelial",
  '17' = "Naive_Bcell",
  '18' = "Granulocyte",
  '19' = "CD31+",
  '20' = "PDPN+",
  '21' = "Epithelial",
  '22' = "Macrophage",
  '23' = "DC",
  '24' = "DC",
  '25' = "Granulocyte",
  '26' = "Epithelial",
  '27' = "Granulocyte",
  '28' = "Epithelial",
  '29' = "Naive_Bcell"
)
```

```{r}
parent_cluster_dict <- data.frame(
  "cluster" = names(broad_cluster_ls),
  "broad" = unlist(broad_cluster_ls)
)

immune_cluster_dict <- data.frame(
  "cluster" = names(immune_broad_cluster_ls),
  "broad" = unlist(immune_broad_cluster_ls),
  "specific" = unlist(immune_specific_cluster_ls)
)
```

```{r}
# parse sample name into meta features
sample_id_split <- strsplit(cell_meta$ImageId, "_")
cell_meta$timepoint <- sapply(sample_id_split, function(x) x[[1]])
cell_meta$slide_id <- sapply(sample_id_split, function(x) x[[2]])
cell_meta$ROI <- sapply(sample_id_split, function(x) x[[3]])
cell_meta$region <- sapply(sample_id_split, function(x) x[[4]])

```

```{r}
cell_meta$cell_cluster <- plyr::mapvalues(
  x = cell_meta$cluster, 
  from = parent_cluster_dict$cluster, to = parent_cluster_dict$broad
)

cell_meta$immune_cluster_broad <- plyr::mapvalues(
  x = cell_meta$immune_cluster, 
  from = immune_cluster_dict$cluster, to = immune_cluster_dict$broad
)

cell_meta$immune_cluster_specific <- plyr::mapvalues(
  x = cell_meta$immune_cluster, 
  from = immune_cluster_dict$cluster, to = immune_cluster_dict$specific
)
```

```{r umap analysis}
umap.settings <- umap::umap.defaults
umap.settings$verbose <- TRUE

print(umap.settings)

set.seed(1234)
umap_res <- umap(imc_norm, config = umap.settings, method = "naive")

umap_df <- data.frame(
  UMAP_1 = umap_res$layout[, 1],
  UMAP_2 = umap_res$layout[, 2]
)
```

```{r}
# merge expression data
cell_meta_umap <- cbind(cell_meta, umap_df)
imc <- cbind(cell_meta_umap, imc_arcsin)
```

```{r}
umap_class_plot <- function(df, group.by, pal = NULL){
  if(!is(df[[group.by]], "factor")) {
    df[[group.by]] <- as.factor(df[[group.by]])
  }
  group_lvls <- levels(df[[group.by]])
  if(is.null(pal)){
    pal <- setNames(
      object = scales::hue_pal()(length(group_lvls)),
      nm = group_lvls
    )
  }
  
  pl <- ggplot(
    data = df, 
    aes(x = UMAP_1, y = UMAP_2, color = .data[[group.by]])
  ) + 
    # cells with NA identifier are in the background
    geom_point(
      data = df[is.na(df[[group.by]]),],
      color = "#CCCCCC",
      size = 0.1
    ) +
    # cells with real labels render in the foreground
    geom_point(
      data = df[!is.na(df[[group.by]]),],
      size = 0.3
    ) +
    theme_minimal() +
    scale_color_manual(values = pal, na.value = "#CCCCCC")
  
  return(pl)
}

umap_feature_plot <- function(df, feature){
  pl <- ggplot(
    data = df, 
    aes(x = UMAP_1, y = UMAP_2, color = .data[[feature]])
  ) + 
    geom_point(
      size = 0.1
    ) +
    theme_minimal() +
    scale_color_gradientn(
      colors = scales::viridis_pal(option = "magma")(10)
    )
  
  return(pl)
}

umap_class_plot(imc, group.by = "cluster")

sum(imc$UMAP_2 < -15)
DT::datatable(
  imc[imc$UMAP_2 < -15,]
)

imc_plot <- imc[imc$UMAP_2 > -15,]

class_colnames <- c(
  "cluster", "immune_cluster", "cell_cluster", 
  "immune_cluster_broad", "immune_cluster_specific", 
  "timepoint", "slide_id", "timepoint", "ROI", "region"
)

feature_colnames <- c(
  "Arginase1", "B220", "BCL6_1", "BCL6_2", "CD10", "CD11c", "CD21", "CD31",
  "CD3e", "CD45", "CD4", "CD68", "CD79a", "CD86", "CD8a", "ECAD", "FOXP3",
  "GZMB", "IgM", "Ly6G", "MHCII", "MMR", "PD1", "PDL1", "PDPN", "Pankeratin",
  "S100A9", "TCF1", "TOX", "Vimentin", "aSMA", "IgD"
)

for(i in seq_along(class_colnames)) {
  cl <- class_colnames[i]
  pl <- umap_class_plot(imc_plot, group.by = cl)
  print(pl)
  ggsave(
    plot = pl,
    filename = paste0(figure_dir, "/", "umap_", "classifier", "_", cl, ".png"),
    width = unit(6, "in"), height = unit(6, "in"), dpi = 300
  )
  
  ggsave(
    plot = pl + theme(legend.position = "none"),
    filename = paste0(figure_dir, "/", "umap_", "classifier", "_", cl, "_noLegend", ".png"),
    width = unit(6, "in"), height = unit(6, "in"), dpi = 300
  )
}

for(i in seq_along(feature_colnames)) {
  fe <- feature_colnames[i]
  pl <- umap_feature_plot(imc_plot, feature = fe)
  print(pl)
  ggsave(
    plot = pl,
    filename = paste0(figure_dir, "/", "umap_", "feature", "_", fe, ".png"),
    width = unit(6, "in"), height = unit(6, "in"), dpi = 300
  )
  
  ggsave(
    plot = pl + theme(legend.position = "none"),
    filename = paste0(figure_dir, "/", "umap_", "feature", "_", fe, "_noLegend", ".png"),
    width = unit(6, "in"), height = unit(6, "in"), dpi = 300
  )
}
```

```{r}
# mean expression heatmap using the immune cell labels
features <- colnames(imc_arcsin)

immune_imc <- imc[!is.na(imc$immune_cluster),]

cluster_mean <- immune_imc[,c("immune_cluster_specific", features)] %>%
  group_by(immune_cluster_specific) %>% 
  dplyr::summarize_all(list(mean))

cluster_mean_mat <- as.matrix(cluster_mean[, features])

rownames(cluster_mean_mat) <- 1:nrow(cluster_mean_mat)

cluster_scaled <- scale(cluster_mean_mat)

rownames(cluster_scaled) <- rownames(cluster_mean)
```

```{r}
n_levels <- nlevels(as.factor(immune_imc$immune_cluster_specific))
ditto_cols <- dittoSeq::dittoColors(1)

# colors for the heatmap
legend_breaks <- seq(from = 0, to = 1, by = 0.2)
colorassigned <- setNames(
  ditto_cols[seq_along(unique(cluster_mean$immune_cluster_specific))],
  sort(unique(cluster_mean$immune_cluster_specific))
)
color_list <- list(clusters=colorassigned)
color_list_byoriginal <- colorassigned[match((cluster_mean$immune_cluster_specific),names(colorassigned))]

rAbar <- rowAnnotation(
  clusters=cluster_mean$immune_cluster_specific,
  col=color_list, gp = gpar(col = "white", lwd = .5),
  counts= anno_barplot(
    as.vector(table(immune_imc$immune_cluster_specific)), gp = gpar(fill=colorassigned),
    border = F, bar_width = 0.75, width = unit(2,"cm")
  )
)

pdf(paste0(figure_dir, "/", "immune_cluster_specific_clusterheatmap.pdf"),width=10,height=8)
Heatmap(
  cluster_scaled,
  column_title="Phenograph Clusters", name = "scaled",
  col=rev(brewer.rdbu(100)), cluster_columns = T, cluster_rows = T,
  border = NA, rect_gp = gpar(col = "white", lwd = .5), right_annotation = rAbar,
  show_row_names = T, row_names_gp = gpar(fontsize=7), column_names_gp = gpar(fontsize=10),
  heatmap_legend_param = list(at=seq(from = round(min(cluster_scaled)), to = round(max(cluster_scaled)))),
  width = ncol(cluster_scaled)*unit(4, "mm"), height = nrow(cluster_scaled)*unit(4, "mm")
)
dev.off() 
```

```{r}
# export data

# arcsin-transformed expression with meta data integrated
write.csv(
  imc, file = paste0(result_dir, "/", "IMC_", "integrated_metadata_arcsinh", ".csv")
)
# with the UMAP outliers excluded
write.csv(
  imc, file = paste0(result_dir, "/", "IMC_", "integrated_metadata_arcsinh", "_outlierExcluded", ".csv")
)

# metadata alone
write.csv(
  cell_meta_umap, file = paste0(result_dir, "/", "IMC_", "metadata", ".csv")
)

# arcsin-transformed expression
write.csv(
  imc_arcsin, file = paste0(result_dir, "/", "IMC_", "arcsinh", ".csv")
)

# normalized (0-1 scale) imc data
write.csv(
  imc_norm, file = paste0(result_dir, "/", "IMC_", "normalized", ".csv")
)

# complete result from UMAP calculation
saveRDS(
  umap_res, file = paste0(result_dir, "/", "UMAP_result", ".rds")
)
```
